requirements.txt
djangorestframework
djangorestframework-simplejwt
django-allauth
django-rest-auth
dj-rest-auth